refactor(catalog): share reflectivity scan classification with TUI and Python
