+++ python/pyref/io/__init__.py
@@ -21,6 +21,7 @@ FITS Files
 Experiment catalog and discovery
 -------------------------------
 - :func:`~pyref.io.resolve_catalog_path`: Beamtime catalog database path (Rust-aligned).
+- :func:`~pyref.io.classify_reflectivity_scan_type`: Theta vs energy scan classification.
 - :func:`~pyref.io.discover_fits`: Discover FITS files under a root path (flat or recursive).
 - :func:`~pyref.io.parse_fits_stem`: Parse a FITS filename stem into sample_name, tag, scan_number, frame_number.
 - :func:`~pyref.io.build_catalog`: Build a per-file catalog DataFrame (names-only or with headers).
@@ -43,6 +44,7 @@ from pyref.io.experiment_names import (
     scan_view,
 )
 from pyref.io.readers import (
+    classify_reflectivity_scan_type,
     get_image,
     get_image_corrected,
     get_image_filtered,
@@ -60,6 +62,7 @@ from pyref.io.readers import (
 __all__ = [
     "ParsedFitsName",
     "build_catalog",
+    "classify_reflectivity_scan_type",
     "discover_fits",
     "experiment_summary",
     "filter_catalog_paths",
